Created a more direct hierarchy for Corp/Runner ID
Major rehaul of how I was doing identity creation.
Created Runner and Corp types... these will be the main objects that allow
a player to take various actions and track various stats for the player.
I thought it fitting that the identity be the player object and
relegate the actual Player object to handling user information and
connection handling.